Move relevant tests to the new test class
